feat(googlelive-nova2-audiobot-homologation): TASK-2177 — Extract browser voice UI + migrate Nova example to VoiceSession
